requirements.txt
google-api-core==2.11.0
google-cloud-bigquery
python-dotenv